np-47035: remove cristin logic from biobank model. That way if we change integration to other than cristin later on we do not need to change nva biobank model
