+++ cristin-biobank/src/main/java/no/unit/nva/biobank/model/nva/Biobank.java
@@ -1,17 +1,8 @@
 package no.unit.nva.biobank.model.nva;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
-import static java.util.Objects.nonNull;
 import static no.unit.nva.cristin.common.Utils.nonEmptyOrDefault;
-import static no.unit.nva.cristin.model.Constants.CRISTIN_IDENTIFIER_TYPE;
-import static no.unit.nva.cristin.model.Constants.FHI_BIOBANK_REGISTRY;
-import static no.unit.nva.cristin.model.Constants.ORGANIZATION_PATH;
-import static no.unit.nva.cristin.model.Constants.PERSON_PATH_NVA;
-import static no.unit.nva.cristin.model.Constants.TYPE;
-import static no.unit.nva.cristin.model.Constants.VALUE;
-import static no.unit.nva.cristin.model.JsonPropertyNames.BIOBANK_ID;
 import static no.unit.nva.cristin.model.JsonPropertyNames.CONTEXT;
-import static no.unit.nva.utils.UriUtils.getNvaApiId;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -22,18 +13,8 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import no.unit.nva.biobank.model.cristin.CristinBiobank;
-import no.unit.nva.biobank.model.cristin.CristinBiobankMaterial;
 import no.unit.nva.commons.json.JsonSerializable;
-import no.unit.nva.cristin.model.CristinApproval;
-import no.unit.nva.cristin.model.CristinDateInfo;
-import no.unit.nva.cristin.model.CristinExternalSource;
-import no.unit.nva.cristin.model.CristinOrganization;
-import no.unit.nva.cristin.model.CristinUnit;
 import no.unit.nva.model.DateInfo;
 import no.unit.nva.model.ExternalSource;
 import no.unit.nva.model.TypedLabel;
@@ -90,27 +71,6 @@ public class Biobank implements JsonSerializable {
         this.startDate = startDate;
         this.status = status;
         this.storeUntilDate = storeUntilDate;
-    }
-
-    public Biobank(CristinBiobank cristinBiobank) {
-        this(
-            getNvaApiId(cristinBiobank.getCristinBiobankId(), BIOBANK_ID),
-            toCristinIdentifier(cristinBiobank),
-            BiobankType.valueOf(cristinBiobank.getType()),
-            cristinBiobank.getName(),
-            cristinBiobank.getMainLanguage(),
-            cristinBiobank.getStoreUntilDate(),
-            cristinBiobank.getStartDate(),
-            cristinBiobank.getStatus(),
-            toDateInfoOrNull(cristinBiobank.getCreated()),
-            toDateInfoOrNull(cristinBiobank.getLastModified()),
-            toCoordinatingUnit(cristinBiobank.getCoordinatingInstitution()),
-            getNvaApiId(cristinBiobank.getCoordinator().getCristinPersonId(), PERSON_PATH_NVA),
-            toProjectOrNull(cristinBiobank),
-            toExternalSources(cristinBiobank.getExternalSources()),
-            toApprovals(cristinBiobank.getApprovals()),
-            toBiobankMaterials(cristinBiobank.getBiobankMaterials())
-        );
     }
 
     public URI getId() {
@@ -223,61 +183,6 @@ public class Biobank implements JsonSerializable {
     @JacocoGenerated
     public String toString() {
         return this.toJsonString();
-    }
-
-    private static List<Map<String, String>> toCristinIdentifier(CristinBiobank cristinBiobank) {
-
-        var cristinBiobankIdOrNull = nonNull(cristinBiobank.getCristinBiobankId())
-            ? Map.of(TYPE, CRISTIN_IDENTIFIER_TYPE, VALUE, cristinBiobank.getCristinBiobankId())
-            : null;
-
-        var biobankIdOrNull = nonNull(cristinBiobank.getBiobankId())
-            ? Map.of(TYPE, FHI_BIOBANK_REGISTRY, VALUE, cristinBiobank.getBiobankId())
-            : null;
-
-        return Stream.of(cristinBiobankIdOrNull,biobankIdOrNull)
-                   .filter(Objects::nonNull)
-                   .toList();
-    }
-
-
-    private static URI toCoordinatingUnit(CristinOrganization cristinOrganization) {
-        return
-            Optional.ofNullable(cristinOrganization)
-                .map(CristinOrganization::getInstitutionUnit)
-                .map(CristinUnit::getCristinUnitId)
-                .map(unitId -> getNvaApiId(unitId, ORGANIZATION_PATH))
-                .orElse(null);
-    }
-
-    private static Set<ExternalSource> toExternalSources(Set<CristinExternalSource> externalSources) {
-        return
-            externalSources.stream()
-                .map(es -> new ExternalSource(es.getSourceReferenceId(), es.getSourceShortName()))
-                .collect(Collectors.toUnmodifiableSet());
-    }
-
-    private static List<BiobankApproval> toApprovals(List<CristinApproval> approvals) {
-        return
-            approvals.stream()
-                .map(BiobankApproval::new)
-                .toList();
-    }
-
-    private static List<TypedLabel> toBiobankMaterials(List<CristinBiobankMaterial> materials) {
-        return
-            materials.stream()
-                .map(item -> new TypedLabel(item.getMaterialCode(),item.getMaterialName()))
-                .toList();
-    }
-
-    private static AssociatedProject toProjectOrNull(CristinBiobank cristinBiobank) {
-        return nonNull(cristinBiobank.getAssociatedProject())
-            ? new AssociatedProject(cristinBiobank.getAssociatedProject()) : null;
-    }
-
-    private static DateInfo toDateInfoOrNull(CristinDateInfo cristinBiobank) {
-        return nonNull(cristinBiobank) ? cristinBiobank.toDateInfo() : null;
     }
 
 }
